pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "plots-lib"
version = "0.1.0"
description = "Plotly plot lib"
authors = [
    {name = "DEV", email = "dev@email.com"}
]

dependencies = [
    "plotly",
    "pillow"
]
